requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic>=2.10
sse-starlette==1.6.5
starlette==0.27.0
python-dotenv==1.0.0
tenacity==8.2.3
haystack-ai==2.11.0
openai>=1.56.1
requests==2.31.0
openai-agents==0.0.4
google-ai-haystack==5.1.0
